--- storm_analysis/homepage/templates/homepage/about.html
@@ -1,9 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %} About the project {% endblock %}
-
-{% block content %} 
-    <h2 style="text-align: center;">About</h2>
-    <p>Send file to run analyse</p>
-    <p>If you have any questions or suggestions regarding the package, please feel free to contact us.</p>
-{% endblock %}
