--- storm_analysis/homepage/templates/homepage/index.html
@@ -1,6 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-<h1>Witaj na stronie głównej!</h1>
-<p>To jest strona główna naszej aplikacji.</p>
-{% endblock %}
