--- storm_analysis/homepage/tests.py
@@ -1,1 +1,0 @@
-# Create your tests here.
