--- storm_analysis/homepage/views.py
@@ -1,9 +1,0 @@
-from django.shortcuts import render
-
-
-def index(request):
-    return render(request, "homepage/index.html")
-
-
-def about(request):
-    return render(request, "homepage/about.html")
