+++ storm_analysis/sa/templates/sa/about.html
@@ -1,0 +1,11 @@
+{% extends "base.html" %}
+
+{% block title %} About the project {% endblock %}
+
+{% block content %} 
+    <div class="container">
+        <h2 style="text-align: center;">About</h2>
+        <p>Send file to run analyse</p>
+        <p>If you have any questions or suggestions regarding the package, please feel free to contact us.</p>
+    </div>
+{% endblock %}
