+++ storm_analysis/sa/templates/sa/index.html
@@ -1,0 +1,6 @@
+{% extends "base.html" %}
+
+{% block content %}
+    <h1>Witaj na stronie głównej!</h1>
+    <p>To jest strona główna naszej aplikacji.</p>
+{% endblock %}
